feat: 공연정보 컴포넌트 공통컴포넌트로 구현 #128
