+++ src/app/concert/[id]/_shared/components/concert-info/concert-info.tsx
@@ -1,19 +1,15 @@
 'use client';
 
 import Image from 'next/image';
-import Link from 'next/link';
 
+import ConcertDetail from '@/shared/components/concert-detail/concert-detail';
 import Badge from '@/shared/components/ui/badge/badge';
 import {
   TICKET_SITE_URL_MAP,
   TICKET_SITE_LABEL_MAP,
 } from '@/shared/constants/type-mapping';
 import { TicketReservationSite, Concert } from '@/shared/types';
-import {
-  formatDate,
-  calculateDday,
-  getPerformancePeriod,
-} from '@/shared/utils/dates';
+import { calculateDday, getPerformancePeriod } from '@/shared/utils/dates';
 import { getPreOpenInfo, getGeneralOpenInfo } from '@/shared/utils/tickets';
 
 import styles from './concert-info.module.scss';
@@ -70,44 +66,28 @@ const ConcertInfo = ({ concertItem }: ConcertInfoProps) => {
           unoptimized
         />
 
-        {/* 태그 */}
-        <div className={styles.tag_container}>
-          {preOpen && (
-            <Badge variant="type-a">
-              선예매까지 {calculateDday(preOpen.openDate)}
-            </Badge>
-          )}
-          {generalOpen && (
-            <Badge variant="type-c">
-              일반예매까지 {calculateDday(generalOpen.openDate)}
-            </Badge>
-          )}
-        </div>
-
-        {/* 제목 */}
-        <div className={styles.title}>{concertName}</div>
-
-        {/* 공연 상세정보 */}
-        <div className={styles.detail_container}>
-          <div className={styles.detail}>
-            <span className={styles.category}>공연일자</span>
-            <span className={styles.info}>
-              {`${formatDate(startDate)} ~ ${formatDate(endDate)}`}
-            </span>
-          </div>
-
-          <div className={styles.detail}>
-            <span className={styles.category}>공연장</span>
-            <span className={styles.info}>{concertHallName}</span>
-          </div>
-
-          <div className={styles.detail}>
-            <span className={styles.category}>예매처</span>
-            <Link className={styles.info} href={siteUrl}>
-              {siteLabel}
-            </Link>
-          </div>
-        </div>
+        <ConcertDetail
+          concertName={concertName}
+          startDate={startDate}
+          endDate={endDate}
+          concertHallName={concertHallName}
+          siteUrl={siteUrl}
+          siteLabel={siteLabel}
+          badges={
+            <>
+              {preOpen && (
+                <Badge variant="type-a">
+                  선예매까지 {calculateDday(preOpen.openDate)}
+                </Badge>
+              )}
+              {generalOpen && (
+                <Badge variant="type-c">
+                  일반예매까지 {calculateDday(generalOpen.openDate)}
+                </Badge>
+              )}
+            </>
+          }
+        />
       </div>
     </div>
   );
