+++ src/app/concert/form/[id]/_shared/components/form-info/form-info.module.scss
@@ -1,0 +1,7 @@
+@use '@/styles/index' as *;
+
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
