--- src/app/concert/form/[id]/_shared/components/form/form.module.scss
@@ -1,47 +1,0 @@
-@use '@/styles/index' as *;
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: 100%;
-
-  .tag_container {
-    display: flex;
-    gap: 8px;
-  }
-
-  .title {
-    @include title-md;
-    margin-top: 8px;
-    color: var(--grayscale-700, #222);
-  }
-
-  .detail_container {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    align-self: stretch;
-    margin-top: 16px;
-    gap: 4px;
-
-    .detail {
-      display: flex;
-      width: 100%;
-      gap: 12px;
-      @include text-md;
-
-      .category {
-        flex: 1;
-        color: var(--grayscale-500);
-      }
-
-      .info {
-        display: flex;
-        justify-content: flex-start;
-        flex: 5;
-        color: var(--grayscale-700);
-      }
-    }
-  }
-}
