+++ src/shared/components/concert-detail/concert-detail.module.scss
@@ -1,0 +1,40 @@
+@use '@/styles/index' as *;
+
+.tag_container {
+  display: flex;
+  gap: 8px;
+}
+
+.title {
+  @include title-md;
+  margin-top: 8px;
+  color: var(--grayscale-700, #222);
+}
+
+.detail_container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  align-self: stretch;
+  margin-top: 16px;
+  gap: 4px;
+
+  .detail {
+    display: flex;
+    width: 100%;
+    gap: 12px;
+    @include text-md;
+
+    .category {
+      color: var(--grayscale-500);
+      flex: 1;
+    }
+
+    .info {
+      display: flex;
+      justify-content: flex-start;
+      flex: 5;
+      color: var(--grayscale-700);
+    }
+  }
+}
